3.1/products/requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
psycopg==3.2.9
sqlparse==0.5.3
typing_extensions==4.14.1
